Add alignment reducer
