+++ packages/noya-state/src/reducers/applicationReducer.ts
@@ -1,35 +1,34 @@
 import Sketch from '@sketch-hq/sketch-file-format-ts';
 import produce from 'immer';
 import { WritableDraft } from 'immer/dist/internal';
+import { AffineTransform, normalizeRect } from 'noya-geometry';
 import { Primitives, SimpleTextDecoration, uuid } from 'noya-renderer';
 import { resizeRect } from 'noya-renderer/src/primitives';
 import { SketchFile } from 'noya-sketch-file';
-import { sum, getIncrementedName, delimitedPath } from 'noya-utils';
+import { delimitedPath, getIncrementedName } from 'noya-utils';
 import { IndexPath } from 'tree-visit';
-import { transformRect, createBounds, normalizeRect } from 'noya-geometry';
 import * as Layers from '../layers';
 import * as Models from '../models';
 import {
   EncodedPageMetadata,
+  findPageLayerIndexPaths,
   getBoundingRect,
+  getCurrentComponentsTab,
   getCurrentPage,
   getCurrentPageIndex,
   getCurrentPageMetadata,
+  getCurrentTab,
   getLayerRotation,
   getLayerRotationMultiplier,
-  getLayerTransformAtIndexPath,
   getSelectedLayerIndexPaths,
   getSelectedLayerIndexPathsExcludingDescendants,
-  getSelectedRect,
-  getCurrentTab,
-  findPageLayerIndexPaths,
-  visitStyleColors,
-  visitLayerColors,
   groupThemeComponents,
-  getCurrentComponentsTab,
+  visitLayerColors,
+  visitStyleColors,
 } from '../selectors/selectors';
-import { Bounds, Point, UUID } from '../types';
-import { AffineTransform } from 'noya-geometry';
+import { Point, UUID } from '../types';
+import { SelectionType, updateSelection } from '../utils/selection';
+import { AlignmentAction, alignmentReducer } from './alignmentReducer';
 import {
   CompassDirection,
   createInitialInteractionState,
@@ -37,12 +36,11 @@ import {
   interactionReducer,
   InteractionState,
 } from './interactionReducer';
-import { SetNumberMode, StyleAction, styleReducer } from './styleReducer';
 import {
   StringAttributeAction,
   stringAttributeReducer,
 } from './stringAttributeReducer';
-import { SelectionType, updateSelection } from '../utils/selection';
+import { SetNumberMode, StyleAction, styleReducer } from './styleReducer';
 
 export type { SetNumberMode };
 
@@ -94,17 +92,6 @@ export type Action =
   | [type: 'deletePage']
   | [type: 'renamePage', name: string]
   | [type: 'duplicatePage']
-  | [type: 'distributeLayers', placement: 'horizontal' | 'vertical']
-  | [
-      type: 'alignLayers',
-      placement:
-        | 'left'
-        | 'centerHorizontal'
-        | 'right'
-        | 'top'
-        | 'centerVertical'
-        | 'bottom',
-    ]
   | [type: 'setLayerRotation', rotation: number, mode?: SetNumberMode]
   | [type: 'setFixedRadius', amount: number, mode?: SetNumberMode]
   | [type: 'setSwatchColor', swatchId: string | string[], color: Sketch.Color]
@@ -150,11 +137,12 @@ export type Action =
     ]
   | StyleAction
   | StringAttributeAction
+  | AlignmentAction
   | [type: 'setTextAlignment', value: number]
   | [type: 'setTextDecoration', value: SimpleTextDecoration]
   | [type: 'setTextTransform', value: number];
 
-export function reducer(
+export function applicationReducer(
   state: ApplicationState,
   action: Action,
 ): ApplicationState {
@@ -379,155 +367,9 @@ export function reducer(
         draft.selectedPage = pages[newIndex].do_objectID;
       });
     }
-    case 'distributeLayers': {
-      const page = getCurrentPage(state);
-      const pageIndex = getCurrentPageIndex(state);
-      const layerIndexPaths = getSelectedLayerIndexPaths(state);
-      const selectedRect = getSelectedRect(state);
-      const [, axis] = action;
-
-      return produce(state, (draft) => {
-        const layers = accessPageLayers(draft, pageIndex, layerIndexPaths);
-        const combinedWidths = sum(layers.map((layer) => layer.frame.width));
-        const combinedHeights = sum(layers.map((layer) => layer.frame.height));
-        const differenceWidth = selectedRect.width - combinedWidths;
-        const differenceHeight = selectedRect.height - combinedHeights;
-        const gapX = differenceWidth / (layers.length - 1);
-        const gapY = differenceHeight / (layers.length - 1);
-        const sortBy = axis === 'horizontal' ? 'midX' : 'midY';
-
-        // Bounds are all transformed to the page's coordinate system
-        function getNormalizedBounds(
-          page: Sketch.Page,
-          layerIndexPath: IndexPath,
-        ): Bounds {
-          const layer = Layers.access(page, layerIndexPath);
-          const transform = getLayerTransformAtIndexPath(
-            page,
-            layerIndexPath,
-            AffineTransform.identity,
-          );
-          return createBounds(transformRect(layer.frame, transform));
-        }
-
-        const sortedLayerIndexPaths = layerIndexPaths.sort(
-          (a, b) =>
-            getNormalizedBounds(page, a)[sortBy] -
-            getNormalizedBounds(page, b)[sortBy],
-        );
-
-        let currentX = 0;
-        let currentY = 0;
-
-        sortedLayerIndexPaths.forEach((layerIndexPath) => {
-          const transform = getLayerTransformAtIndexPath(
-            page,
-            layerIndexPath,
-            AffineTransform.identity,
-          ).invert();
-          const layer = Layers.access(
-            draft.sketch.pages[pageIndex], // access page again since we need to write to it
-            layerIndexPath,
-          );
-
-          switch (axis) {
-            case 'horizontal': {
-              const newOrigin = transform.applyTo({
-                x: selectedRect.x + currentX,
-                y: 0,
-              });
-              currentX += layer.frame.width + gapX;
-              layer.frame.x = newOrigin.x;
-              break;
-            }
-            case 'vertical': {
-              const newOrigin = transform.applyTo({
-                x: 0,
-                y: selectedRect.y + currentY,
-              });
-              currentY += layer.frame.height + gapY;
-              layer.frame.y = newOrigin.y;
-              break;
-            }
-          }
-        });
-      });
-    }
+    case 'distributeLayers':
     case 'alignLayers': {
-      const page = getCurrentPage(state);
-      const pageIndex = getCurrentPageIndex(state);
-      const layerIndexPaths = getSelectedLayerIndexPaths(state);
-      const selectedRect = getSelectedRect(state);
-      const [, placement] = action;
-
-      return produce(state, (draft) => {
-        const selectedBounds = createBounds(selectedRect);
-        const midX = selectedRect.x + selectedRect.width / 2;
-        const midY = selectedRect.y + selectedRect.height / 2;
-
-        layerIndexPaths.forEach((layerIndexPath) => {
-          const transform = getLayerTransformAtIndexPath(
-            page,
-            layerIndexPath,
-            AffineTransform.identity,
-          ).invert();
-          const layer = Layers.access(
-            draft.sketch.pages[pageIndex], // access page again since we need to write to it
-            layerIndexPath,
-          );
-
-          switch (placement) {
-            case 'left': {
-              const newOrigin = transform.applyTo({
-                x: selectedBounds.minX,
-                y: 0,
-              });
-              layer.frame.x = newOrigin.x;
-              break;
-            }
-            case 'centerHorizontal': {
-              const newOrigin = transform.applyTo({
-                x: midX - layer.frame.width / 2,
-                y: 0,
-              });
-              layer.frame.x = newOrigin.x;
-              break;
-            }
-            case 'right': {
-              const newOrigin = transform.applyTo({
-                x: selectedBounds.maxX - layer.frame.width,
-                y: 0,
-              });
-              layer.frame.x = newOrigin.x;
-              break;
-            }
-            case 'top': {
-              const newOrigin = transform.applyTo({
-                x: 0,
-                y: selectedBounds.minY,
-              });
-              layer.frame.y = newOrigin.y;
-              break;
-            }
-            case 'centerVertical': {
-              const newOrigin = transform.applyTo({
-                x: 0,
-                y: midY - layer.frame.height / 2,
-              });
-              layer.frame.y = newOrigin.y;
-              break;
-            }
-            case 'bottom': {
-              const newOrigin = transform.applyTo({
-                x: 0,
-                y: selectedBounds.maxY - layer.frame.height,
-              });
-              layer.frame.y = newOrigin.y;
-              break;
-            }
-          }
-        });
-      });
+      return alignmentReducer(state, action);
     }
     case 'addNewBorder':
     case 'addNewFill':
@@ -1469,7 +1311,7 @@ export function reducer(
  * Immer will duplicate any objects we access within a produce method, so we
  * don't want to walk every layer, since that would duplicate all of them.
  */
-function accessPageLayers(
+export function accessPageLayers(
   state: WritableDraft<ApplicationState>,
   pageIndex: number,
   layerIndexPaths: IndexPath[],
