+++ packages/noya-state/src/reducers/historyReducer.ts
@@ -3,7 +3,7 @@ import produce from 'immer';
 import {
   ApplicationState,
   Action,
-  reducer,
+  applicationReducer,
   createInitialState,
 } from './applicationReducer';
 import { createInitialInteractionState } from './interactionReducer';
@@ -56,7 +56,7 @@ export function historyReducer(state: HistoryState, action: HistoryAction) {
         });
       }
     default:
-      const nextState = reducer(currentState, action);
+      const nextState = applicationReducer(currentState, action);
       const mergableEntry = getMergableHistoryEntry(state, action[0]);
       const sketchFileChanged = currentState.sketch !== nextState.sketch;
       return produce(state, (draft) => {
